added statuschecker to rabbitmq provider and resolved issues with rspec test execution
