+++ eventq_rabbitmq/lib/eventq_rabbitmq/version.rb
@@ -1,3 +1,3 @@
 module EventqRabbitmq
-  VERSION = "1.12.0"
+  VERSION = "1.13.0"
 end
