+++ eventq_rabbitmq/script/test.sh
@@ -3,4 +3,4 @@
 echo start rspec tests
 docker-compose up -d
 
-docker exec -it gem_test_runner bash -c "cd gem_src && sleep 2 && bundle install && bundle exec rspec $*"
+docker exec -it gem_test_runner bash -c "cd gem_src && sleep 4 && bundle install && bundle exec rspec $*"
