+++ eventq_rabbitmq/spec/eventq_rabbitmq/rabbitmq_eventq_client_spec.rb
