+++ eventq_rabbitmq/spec/eventq_rabbitmq/rabbitmq_queue_client_spec.rb
